Do this the most evil way possible.  This appear to work in my tests. libgmp fellas are being unfortunately uncooperative in improving their software, so it make sense to have a tenable solution without them.
This can and will badly malfunction at some point, I'm not sure I ever
want to run this in production -- time will tell.

This uses signals catching the SIGABRT and SIGFPE (the two raised in
libgmp) and siglongjmps out of them.  In single threaded programs the
SIGFPE should be entirely safe, the SIGABRT is a bit more risky.  We
limit the siglongjmp to when we know we've called into libgmp.

And I will end this with a parting note: libgmp can suck it.